+++ src/main/java/com/ourmenu/backend/domain/user/application/AsyncEmailSenderService.java
@@ -1,0 +1,29 @@
+package com.ourmenu.backend.domain.user.application;
+
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class AsyncEmailSenderService {
+
+    private final JavaMailSender emailSender;
+
+    @Async("mailExecutor")
+    public void sendEmail(String toEmail, String title, String content) throws MessagingException {
+        MimeMessage message = emailSender.createMimeMessage();
+        MimeMessageHelper helper = new MimeMessageHelper(message, true);
+        helper.setTo(toEmail);
+        helper.setSubject(title);
+        helper.setText(content, true);
+        helper.setReplyTo("ourmenuv2@gmail.com");
+        emailSender.send(message);
+    }
+}
