+++ src/main/java/com/ourmenu/backend/domain/user/application/EmailService.java
@@ -9,11 +9,8 @@ import com.ourmenu.backend.domain.user.exception.ConfirmCodeNotFoundException;
 import com.ourmenu.backend.domain.user.exception.NotMatchConfirmCodeException;
 import com.ourmenu.backend.domain.user.exception.SendCodeFailureException;
 import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
 
 import java.util.concurrent.ThreadLocalRandom;
@@ -23,19 +20,9 @@ import java.util.concurrent.ThreadLocalRandom;
 @RequiredArgsConstructor
 public class EmailService {
 
-    private final JavaMailSender emailSender;
+    private final AsyncEmailSenderService asyncEmailSenderService;
     private final int CONFIRM_CODE_LENGTH = 6;
     private final ConfirmCodeRepository confirmCodeRepository;
-
-    public void sendEmail(String toEmail, String title, String content) throws MessagingException {
-        MimeMessage message = emailSender.createMimeMessage();
-        MimeMessageHelper helper = new MimeMessageHelper(message, true);
-        helper.setTo(toEmail);
-        helper.setSubject(title);
-        helper.setText(content, true);
-        helper.setReplyTo("ourmenuv2@gmail.com");
-        emailSender.send(message);
-    }
 
     public EmailResponse sendCodeToEmail(EmailRequest request){
         String email = request.getEmail();
@@ -48,9 +35,8 @@ public class EmailService {
                 + "</body>"
                 + "</html>";
 
-
         try {
-            sendEmail(email, title, content);
+            asyncEmailSenderService.sendEmail(email, title, content);
         }catch (MessagingException e){
             throw new SendCodeFailureException();
         }
